# Conflicts:
#	pom.xml
#	src/main/java/kg/attractor/jobsearch/controller/ApplicantController.java
#	src/main/java/kg/attractor/jobsearch/controller/EmployerController.java
#	src/main/java/kg/attractor/jobsearch/dao/UserDao.java
#	src/main/java/kg/attractor/jobsearch/dto/UserDto.java
#	src/main/java/kg/attractor/jobsearch/dto/VacancyDto.java
#	src/main/java/kg/attractor/jobsearch/exception/ErrorResponseBody.java
#	src/main/java/kg/attractor/jobsearch/service/ErrorService.java
#	src/main/java/kg/attractor/jobsearch/service/ResumeService.java
#	src/main/java/kg/attractor/jobsearch/service/UserService.java
#	src/main/java/kg/attractor/jobsearch/service/impl/ErrorServiceImpl.java
#	src/main/java/kg/attractor/jobsearch/service/impl/ResumeServiceImpl.java
#	src/main/java/kg/attractor/jobsearch/service/impl/UserServiceImpl.java
#	src/main/java/kg/attractor/jobsearch/service/impl/VacancyServiceImpl.java

+++ src/main/java/kg/attractor/jobsearch/service/VacancyService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface VacancyService {
     List<VacancyDto> getVacancies();
 
-    VacancyDto getVacancyById(long id);
+    VacancyDto getVacancyById(Integer id);
 
     void createVacancy(VacancyDto vacancyDto);
 
@@ -17,7 +17,7 @@ public interface VacancyService {
 
     List<VacancyDto> getVacanciesByCategoryId(Integer categoryId);
 
-    void editVacancy(Long id, VacancyDto vacancyDto);
+    void editVacancy(Integer id, VacancyDto vacancyDto);
 
     List<VacancyDto> getVacanciesByCategory(String category);
 }
